+++ NEWS.md
@@ -6,6 +6,7 @@
 * Add functions for reordering factors (such as for ggplot2 bar plots) thanks to @tmastny (#110).
 * Update to `tibble()` where appropriate, thanks to @luisdza (#136).
 * Clarify documentation about impact of lowercase conversion on URLs (#139).
+* Change how sentiment lexicons are accessed from package (remove NRC lexicon entirely, access AFINN and Loughran lexicons via textdata package so they are no longer included in this package).
 
 # tidytext 0.2.0